Refactor conversation handlers into modular package structure
