+++ handlers/__init__.py
@@ -1,0 +1,2 @@
+# Aggregates handlers and shared state
+
