+++ handlers/github/__init__.py
@@ -1,0 +1,2 @@
+# Subpackage for GitHub-related handlers
+
