+++ handlers/github/menu.py
@@ -1,0 +1,4 @@
+from github_menu_handler import GitHubMenuHandler  # temporary re-export stub for compatibility
+
+__all__ = ["GitHubMenuHandler"]
+
