+++ handlers/states.py
@@ -1,0 +1,5 @@
+# Shared conversation states
+
+# Keep the same numeric order as before for compatibility
+GET_CODE, GET_FILENAME, GET_NOTE, EDIT_CODE, EDIT_NAME = range(5)
+
